WuKong--优化秒杀业务逻辑，预减库存，支付失败或者超时未支付回补库存。支付成功真正扣减库存

+++ wukong-provider/src/main/java/com/wukong/provider/service/impl/DubboOrderServiceImpl.java
@@ -21,7 +21,7 @@ public class DubboOrderServiceImpl implements DubboOrderService {
     public BaseResult addOrder(GoodsVO goodsVO, String username) {
 
         System.out.println("全局事务id ：" + RootContext.getXID());
-        int num = orderService.createOrder(goodsVO, username);
+        Long num = orderService.createOrder(goodsVO, username);
         if(num > 0){
             return BaseResult.success(num);
         } else {
